Big refactor.
This is a big squash of several commits that were made during an experimental
phase. Since lots of the earlier commits were obsolted by the latter ones, and
since this gem is not in production anywhere, and still very young, I decided to
squash everything, and put in a pretty big commit with a lot of meat as to how
the ingest process is intended to work.

Instead of detailing the incremental steps of how I got to this point, I'll put
forth the main idea:

* The SIP object is a model for reading files on disk that are to be ingested.

* The Ingester object has two responsibilities:
  1. Read in the mapping configuration (a YAML file).
  1. Initiate the ingestion process

* The mapping file (once parsed) specifies one or more map operations.

* Each map operation consists of a fetch operation, followed by an assignment
  operation.

* MapOperation, Fetcher, and Assigner classes are used for modeling map, fetch,
  and assignment operations, respectively.

* The rules by which values are fetched from the SIP, and then assigned to
  another object, are specified in the mapping configuration.

That's the basic gist of how the ingest process is designed. The state of this
commit is work-in-progress, but it's time to get said progress synced to Github,
so that's what we're doing here.